app/scraper/scraping_job.py
from typing import Dict, Optional
import asyncio

class ScrapingJob:
    def __init__(self, data: Dict, config: Optional[Dict] = None):
        if 'payload' in data:
            self.data = data.get('payload')
        else:
            self.data = data
        self.multiple_pages = self.data.get('multiple_pages', False)
        self.nested = self.data.get('nested', False)
        self.urls = self.data.get('urls')
        self.actions = self.data.get('actions')

        if self.multiple_pages:
            self.next_page_button_xpath = self.data.get('next_page_button_xpath', None)

        if self.nested:
            if config is not None:
                self.config = config
            else:
                raise ValueError("Config is required for nested scraping.")
        else:
            self.config = self.data.get('config', None)

    async def initialize_browser(self):
        """
        Initialize a Puppeteer browser based on the provided configuration.

        Returns:
        - None
        """
        if not hasattr(self, 'browser') or self.browser is None:
            browser_options = {"headless": self.config.get("headless", True)}
            self.browser = await launch(options=browser_options)

    async def perform_actions(self):
        """
        Perform scraping actions based on the configuration for each URL.

        Returns:
        - None
        """
        await self.initialize_browser()

        # Iterate over each URL and perform actions asynchronously
        tasks = [self.perform_actions_for_url(url) for url in self.urls]
        await asyncio.gather(*tasks)

    async def perform_actions_for_url(self, url: str):
        """
        Perform scraping actions for a specific URL.

        Parameters:
        - url (str): The URL to perform actions on.

        Returns:
        - None
        """
        # Your logic to set up the page for the given URL
        page = await self.browser.newPage()
        await page.goto(url)

        # Perform actions for the given URL
        for action in self.actions:
            if action['type'] == 'job':
                nested_job = ScrapingJob(action, config=self.config)
                await nested_job.perform_actions()
            elif action['type'] == 'click':
                await self.click(page, action['xpath'])
            elif action['type'] == 'scrape':
                await self.scrape(page, action['xpath'], action['label'])

        # Close the page after performing actions
        await page.close()

    async def click(self, xpath: str):
        # Implement your logic to simulate a click action
        page = await self.browser.newPage()
        await page.goto(self.config.get("browser", "about:blank"))
        await page.click(xpath)
        await page.close()

    async def scrape(self, xpath: str, label: str):
        # Implement your logic to perform scraping based on the provided XPath
        # Use the 'label' to identify or label the scraped payload
        page = await self.browser.newPage()
        await page.goto(self.config.get("browser", "about:blank"))
        # Perform scraping logic
        await page.close()
